fix: large file download via public share links and hash worker
Stream file chunks lazily (one at a time) instead of opening all file
handles simultaneously, which exceeded OS file descriptor limits for
files with many chunks (~2500+ for 10GB). Also refactored S3 provider
to fetch chunks lazily instead of loading all into memory.

Hash worker now reads files in 4MB slices with an incremental SHA-256
implementation instead of loading the entire file into a single
ArrayBuffer, which failed for large files.

Secondary fixes:
- Content-Length header consistency in links download
- Read-only file opens for download paths
- Chunk query param type i32 -> i64 in storage HEAD handler

Closes #154

+++ storage/src/routes/download.rs
@@ -78,7 +78,7 @@ pub(crate) async fn head(
     let file_id: String = util::actix::path_var(&req, "file_id")?;
     let file_id = Uuid::from_str(&file_id)?;
     claims.validate_transfer_path(file_id, "download")?;
-    let chunk = util::actix::query_var::<i32>(&req, "chunk").ok();
+    let chunk = util::actix::query_var::<i64>(&req, "chunk").ok();
     let format = util::actix::query_var::<String>(&req, "format").ok();
 
     let file = get_file(&context, claims.sub(), file_id)
